Improved mod suggestion message handling code to use db and fixed lint errors

+++ discord-bot/src/events/messageUpdated.ts
@@ -5,15 +5,24 @@ import { Events, Message } from "discord.js";
 export default {
   name: Events.MessageUpdate,
   once: false,
-  async execute(_oldMessage: Message, newMessage: Message) {
+  async execute(oldMessage: Message, newMessage: Message) {
     if (
       newMessage.channelId == bot.settings.modSuggestionChannelId &&
       bot.settings.modSuggestionsEnabled
     ) {
-      await newMessage.delete();
-      logger.info(
-        `${newMessage.author.displayName}(${newMessage.author.id}) editted a message(${newMessage.id}) in mod_suggestions(${newMessage.channelId}) and it was deleted.`
-      );
+      if (oldMessage.content == newMessage.content) return;
+
+      const msg = await bot.mod_suggestion_msgs_database.findOne({
+        where: { id: newMessage.id }
+      });
+
+      if (msg != null) {
+        await newMessage.delete();
+        logger.info(
+          `${newMessage.author.displayName}(${newMessage.author.id}) editted a message(${newMessage.id}) in mod_suggestions(${newMessage.channelId}) and it was deleted.`
+        );
+      }
+
       return;
     }
   }
